chunk: Fix writing line info in the old path
Accidentally changed the order when run-length encoding was implemented.

chunk->count needs to be incremented as the first thing in the new path since
the function can now return early if the passed in line matches previous line.
Count still needs to be incremented in that case.

This is not so for the old path, where each byte in the chunk code array has a
corresponding line info in the lines array, at the same index. Count must be
incremented after storing both of them.

Rename line_start to be more descriptive as well.

Fixes: 64c33f5cbcc355d827b4bf578f487f9473cc1bf2 ("chunk: Implement run-length encoding for line information")

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>